Version 2.1
The extension first check if there are tables in the page, if so assumes
that it is the correct page. And also the check boxes are made to look
like switches. And the README us update to have info about version and
release date, including steps on how to update the extension.